Fix: keep KAN encoders in float32 to prevent NaN after grid update
Root cause of permanent NaN at step 10001+:
  self.encoders.to(device=device, dtype=bfloat16) converted the KANLinear
  grid buffer (B-spline knot positions) to bfloat16. After update_grid,
  adjacent knots in bfloat16 can quantize to the same value, making the
  B-spline basis matrix A degenerate. torch.linalg.lstsq on a degenerate
  matrix produces NaN spline weights → all subsequent forward passes NaN.
  Skipping bad steps did not help because the weights were already corrupted.

Fix:
- encoders.to(device=device) only — dtype stays float32 (grid stays f32)
- KANEncoder.forward: casts input x to float32, casts output back to x.dtype
  so the rest of the network (bfloat16) is unaffected
- update_grid: .float() comment updated (still needed for bfloat16 input)

Memory cost: ~1.3 GB (float32 encoders) vs ~672 MB (bfloat16).
Acceptable — prevents permanent NaN that makes the remaining 40K steps useless.

+++ kan_clt/kan_transcoder.py
@@ -76,7 +76,12 @@ class KANCrossLayerTranscoder(nn.Module):
         self.scan = scan
 
         # KAN encoders — one per layer
-        # KANLinear (efficient_kan) initializes on CPU by default; move to target device/dtype
+        # Keep encoders in float32 regardless of training dtype.
+        # KANLinear stores a `grid` buffer for B-spline knot positions; converting it to
+        # bfloat16 causes adjacent knots to quantize to the same value → degenerate basis
+        # matrix in update_grid's lstsq → NaN spline weights after every grid update.
+        # KANEncoder.forward() casts inputs to float32 and outputs back to model dtype,
+        # so the rest of the network is unaffected.
         self.encoders = nn.ModuleList([
             KANEncoder(
                 d_model=d_model,
@@ -86,7 +91,7 @@ class KANCrossLayerTranscoder(nn.Module):
             )
             for _ in range(n_layers)
         ])
-        self.encoders.to(device=device, dtype=dtype)
+        self.encoders.to(device=device)  # move to device only; dtype stays float32
 
         # Encoder biases (applied after KAN forward, before activation)
         self.b_enc = nn.Parameter(
